+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Estrogen `@VERSION@` `1.20.1` `Fabric & Forge`
 ### Changelog:
-- Fixed an EnderIO specific crash
+- Fixed crashes and bugs
+- Added the ability to enchant patches with capacity